add login, logout, signup

+++ model/contacts.js
@@ -2,24 +2,44 @@ const Contacts = require('./schemas/contacts')
 // with db
 
 
-const listContacts = async () => {
-  const results = await Contacts.find()
+const listContacts = async (userId, query) => {
+  const { sortBy, sortByDesc, filter, favorite = null, limit = 5, offset = 0} = query
+  const optionsSearch = { owner: userId }
+  if (favorite !== null) {
+    optionsSearch.favorite = favorite
+  }
+  const results = await Contacts.paginate(optionsSearch, {
+    limit,
+    offset,
+    sort: {
+      ...(sortBy ? { [`${sortBy}`]: 1 } : {}),
+      ...(sortByDesc ? { [`${sortByDesc}`]: -1 } : {}),
+    },
+    select: filter ? filter.split('|').join(' ') : '',
+    populate: {
+      path: 'owner',
+      select: 'name emeil gender phone'
+    },
+  })  
   return results
 }
 
-const getContactById = async (id) => {
-  const result = await Contacts.findOne({ _id: id })
+const getContactById = async (userId, id) => {
+  const result = await Contacts.findOne({ _id: id, owner: userId }).populate({
+    path: 'owner',
+    select: 'name emeil gender'
+  })
   return result
 }
 
-const removeContact = async (id) => {
-  const result = await Contacts.findByIdAndRemove({ _id: id })
+const removeContact = async (userId, id) => {
+  const result = await Contacts.findByIdAndRemove({ _id: id, owner: userId  })
   return result
 }
 
-const addContact = async (body) => {
+const addContact = async (userId, body) => {
   // try {
-    const results = await Contacts.create(body)
+  const results = await Contacts.create({ ...body, owner: userId })
     return results
   // } catch (e) {
   //   const err = new Error(e)
@@ -30,9 +50,9 @@ const addContact = async (body) => {
   // }
 }
 
-const updateContact = async (id, body ) => {
+const updateContact = async (userId, id, body ) => {
   const result = await Contacts.findByIdAndUpdate(
-    { _id: id },
+    { _id: id, owner: userId },
     { ...body  },
     { new: true},
   )
